teacher page bugs fixed

+++ back-end/controller/studentSectionController.js
@@ -20,8 +20,9 @@ async function postAllClass(req, res) {
 }
 
 async function updateAllClass(req, res) {
+  console.log(req.body);
   try {
-    const updated = await ClassGroup.findByIdAndUpdate(req.params.id, req.body, {
+    const updated =  await ClassGroup.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
     });
     res.json(updated);
@@ -31,9 +32,23 @@ async function updateAllClass(req, res) {
 }
 
 async function deleteAllClass(req, res) {
+  console.log(req.body);
   try {
-    await ClassGroup.findByIdAndDelete(req.params.id);
-    res.json({ message: "Deleted successfully" });
+    const user=await ClassGroup.findById(req.params.id);
+    console.log(user.sections.length);
+    if(user.sections.length>1)
+    {
+      const updated = await ClassGroup.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+      });
+      res.json(updated);
+    }
+    else{
+      await ClassGroup.findByIdAndDelete(req.params.id);
+      res.json({ message: "Deleted successfully" });
+    }
+    
+    
   } catch (err) {
     res.status(400).json({ error: err.message });
   }
